requirements.txt
aioswitcher==1.2.5
dataclasses-json==0.5.6
datadog==0.43.0
freezegun==1.1.0
google-api-python-client==2.33.0
google-assistant-grpc==0.2.1
google-assistant-sdk==0.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
pyowm==3.2.0
python-json-logger==2.0.2